Send smbclient password via pipe instead of command line arg.
Various minor cleanups.

Use pipespawn instead of system() in selfcheck to clean
up quoting problems and be consistent with sendsize and
sendbackup.

Clean up listed incremental file handling.

Backport fullread() and fullwrite() from main branch.

Update and clean up docs/SAMBA.

+++ client-src/sendsize.c
@@ -24,7 +24,7 @@
  * file named AUTHORS, in the root directory of this distribution.
  */
 /* 
- * $Id: sendsize.c,v 1.97.2.13 2000/10/11 02:08:26 martinea Exp $
+ * $Id: sendsize.c,v 1.97.2.13.4.1 2001/03/15 02:18:24 jrjackson Exp $
  *
  * send estimated backup sizes using dump
  */
@@ -32,6 +32,7 @@
 #include "amanda.h"
 #include "pipespawn.h"
 #include "amandates.h"
+#include "util.h"
 #include "getfsent.h"
 #include "version.h"
 
@@ -219,7 +220,7 @@ char **argv;
 	s[-1] = '\0';
 
 	spindle = 0;				/* default spindle */
-	amfree(exclude);				/* default is no exclude */
+	amfree(exclude);			/* default is no exclude */
 
 	skip_whitespace(s, ch);			/* find the spindle */
 	if(ch != '\0') {
@@ -361,7 +362,8 @@ disk_estimates_t *est;
       case 0:
 	break;
       case -1:
-        error("calc_estimates: fork returned: %s", strerror(errno));
+        error("%s: calc_estimates: fork returned: %s",
+	      get_pname(), strerror(errno));
       default:
 	return;
       }
@@ -433,7 +435,7 @@ disk_estimates_t *est;
 
     switch(calcpid = fork()) {
     case -1:
-        error("%s: fork returned: %s", cmd, strerror(errno));
+        error("%s: %s: fork returned: %s", get_pname(), cmd, strerror(errno));
     default:
         break;
     case 0:
@@ -829,14 +831,8 @@ long getsize_dump(disk, level)
 # endif
 #endif
 	{
-	  char *e;
-
-	  e = strerror(errno);
-	  dbprintf(("%s: exec %s failed or no dump program available: %s\n",
-		    get_pname(), cmd, e));
-	  error("%s: exec %s failed or no dump program available: %s",
-		get_pname(), cmd, e);
-	  exit(1);
+	    error("%s: exec %s failed or no dump program available: %s",
+		  get_pname(), cmd, strerror(errno));
 	}
     }
 
@@ -919,27 +915,46 @@ long getsize_dump(disk, level)
 
 #ifdef SAMBA_CLIENT
 long getsize_smbtar(disk, level, exclude_spec)
-     char *disk;
-     int level;
-     char *exclude_spec;
+    char *disk;
+    int level;
+    char *exclude_spec;
 {
-    int pipefd, nullfd, dumppid;
+    int pipefd = -1, nullfd = -1, passwdfd = -1;
+    int dumppid;
     long size;
     FILE *dumpout;
-    char *tarkeys, *sharename, *pass, *domain = NULL;
+    char *tarkeys, *sharename, *user_and_password = NULL, *domain = NULL;
+    int lpass;
+    char *pwtext;
     char *line;
 
-    if ((pass = findpass(disk, &domain)) == NULL) {
-	error("[sendsize : error in smbtar diskline, unable to find password]");
-    }
-    if ((sharename = makesharename(disk, 0)) == NULL) {
-	memset(pass, '\0', strlen(pass));
-	amfree(pass);
+    if ((user_and_password = findpass(disk, &domain)) == NULL) {
 	if(domain) {
 	    memset(domain, '\0', strlen(domain));
 	    amfree(domain);
 	}
-	error("[sendsize : can't make share name of %s]", disk);
+	error("%s-smbtar: cannot find password for %s", get_pname(), disk);
+    }
+    lpass = strlen(user_and_password);
+    if ((pwtext = strchr(user_and_password, '%')) == NULL) {
+	memset(user_and_password, '\0', lpass);
+	amfree(user_and_password);
+	if(domain) {
+	    memset(domain, '\0', strlen(domain));
+	    amfree(domain);
+	}
+	error("%s-smbtar: password field not \'user%%pass\' for %s",
+	      get_pname(), disk);
+    }
+    *pwtext++ = '\0';
+    if ((sharename = makesharename(disk, 0)) == NULL) {
+	memset(user_and_password, '\0', lpass);
+	amfree(user_and_password);
+	if(domain) {
+	    memset(domain, '\0', strlen(domain));
+	    amfree(domain);
+	}
+	error("%s-smbtar: cannot make share name of %s", get_pname(), disk);
     }
     nullfd = open("/dev/null", O_RDWR);
 
@@ -955,26 +970,35 @@ long getsize_smbtar(disk, level, exclude_spec)
 	tarkeys = "archive 1;recurse;dir";
 #endif
 
-    dbprintf(("%s: running \"%s \'%s\' -d %s -U %s -E%s%s -c \'%s\'\"\n",
-	      get_pname(), SAMBA_CLIENT, sharename,
-	      SAMBA_DEBUG_LEVEL, "(see amandapass)", domain ? " -W " : "",
-	      domain ? domain : "", tarkeys));
-
-    dumppid = pipespawn(SAMBA_CLIENT, STDERR_PIPE, &nullfd, &nullfd, &pipefd, 
-	      "smbclient", sharename,
+    dumppid = pipespawn(SAMBA_CLIENT, STDERR_PIPE|PASSWD_PIPE,
+	      &nullfd, &nullfd, &pipefd, 
+	      "PASSWD_FD", &passwdfd,
+	      "smbclient",
+	      sharename,
 	      "-d", SAMBA_DEBUG_LEVEL,
-	      "-U", pass,
+	      "-U", user_and_password,				/* user only */
 	      "-E",
 	      domain ? "-W" : skip_argument,
 	      domain ? domain : skip_argument,
 	      "-c", tarkeys,
-	      (char *)0);
-    memset(pass, '\0', strlen(pass));
-    amfree(pass);
+	      NULL);
     if(domain) {
 	memset(domain, '\0', strlen(domain));
 	amfree(domain);
     }
+    aclose(nullfd);
+    if(fullwrite(passwdfd, pwtext, strlen(pwtext)) < 0) {
+	int save_errno = errno;
+
+	memset(user_and_password, '\0', lpass);
+	amfree(user_and_password);
+	aclose(passwdfd);
+	error("%s-smbtar: password write failed: %s",
+	      get_pname(), strerror(save_errno));
+    }
+    memset(user_and_password, '\0', lpass);
+    amfree(user_and_password);
+    aclose(passwdfd);
     amfree(sharename);
     dumpout = fdopen(pipefd,"r");
 
@@ -1001,8 +1025,8 @@ long getsize_smbtar(disk, level, exclude_spec)
 
     wait(NULL);
 
-    aclose(nullfd);
     afclose(dumpout);
+    pipefd = -1;
 
     return size;
 }
@@ -1015,141 +1039,109 @@ int level;
 char *exclude_spec;
 time_t dumpsince;
 {
-    int pipefd, nullfd, dumppid;
-    long size;
-    FILE *dumpout;
+    int pipefd = -1, nullfd = -1, dumppid;
+    long size = -1;
+    FILE *dumpout = NULL;
     char *incrname = NULL;
+    char *basename = NULL;
     char *dirname = NULL;
+    char *inputname = NULL;
+    FILE *in = NULL;
+    FILE *out = NULL;
     char *exclude_arg = NULL;
     char *efile = NULL, *estr = NULL;
     char *line = NULL;
     char *cmd = NULL;
-    char *cmd_line;
     char dumptimestr[80];
     struct tm *gmtm;
 
 #ifdef GNUTAR_LISTED_INCREMENTAL_DIR
     {
-      char *basename = NULL;
-      char number[NUM_STR_SIZE];
-      char *s;
-      int ch;
+	char number[NUM_STR_SIZE];
+	char *s;
+	int ch;
+	int baselevel;
 
-      basename = vstralloc(GNUTAR_LISTED_INCREMENTAL_DIR,
-			   "/",
-			   host,
-			   disk,
-			   NULL);
-      /*
-       * The loop starts at the first character of the host name,
-       * not the '/'.
-       */
-      s = basename + sizeof(GNUTAR_LISTED_INCREMENTAL_DIR);
-      while((ch = *s++) != '\0') {
-	if(ch == '/' || isspace(ch)) s[-1] = '_';
-      }
-
-      ap_snprintf(number, sizeof(number), "%d", level);
-      incrname = vstralloc(basename, "_", number, ".new", NULL);
-      unlink(incrname);
-
-      if (level == 0) {
-	FILE *out;
-
-notincremental:
-
-	out = fopen(incrname, "w");
-	if (out == NULL) {
-	  dbprintf(("error [opening %s: %s]\n", incrname, strerror(errno)));
-	  amfree(incrname);
-	  amfree(basename);
-	  amfree(dirname);
-	  return -1;
+	basename = vstralloc(GNUTAR_LISTED_INCREMENTAL_DIR,
+			     "/",
+			     host,
+			     disk,
+			     NULL);
+	/*
+	 * The loop starts at the first character of the host name,
+	 * not the '/'.
+	 */
+	s = basename + sizeof(GNUTAR_LISTED_INCREMENTAL_DIR);
+	while((ch = *s++) != '\0') {
+	    if(ch == '/' || isspace(ch)) s[-1] = '_';
 	}
 
-	if (fclose(out) == EOF) {
-	  dbprintf(("error [closing %s: %s]\n", incrname, strerror(errno)));
-	  out = NULL;
-	  amfree(incrname);
-	  amfree(basename);
-	  amfree(dirname);
-	  return -1;
-	}
-	out = NULL;
+	ap_snprintf(number, sizeof(number), "%d", level);
+	incrname = vstralloc(basename, "_", number, ".new", NULL);
+	unlink(incrname);
 
-      } else {
-	FILE *in = NULL, *out;
-	char *inputname = NULL;
-	char buf[BUFSIZ];
-	int baselevel = level;
+	/*
+	 * Open the listed incremental file from the previous level.  Search
+	 * backward until one is found.  If none are found (which will also
+	 * be true for a level 0), arrange to read from /dev/null.
+	 */
+	baselevel = level;
+	while (in == NULL) {
+	    if (--baselevel >= 0) {
+		ap_snprintf(number, sizeof(number), "%d", baselevel);
+		inputname = newvstralloc(inputname,
+					 basename, "_", number, NULL);
+	    } else {
+		inputname = newstralloc(inputname, "/dev/null");
+	    }
+	    if ((in = fopen(inputname, "r")) == NULL) {
+		int save_errno = errno;
 
-	while (in == NULL && --baselevel >= 0) {
-	  ap_snprintf(number, sizeof(number), "%d", baselevel);
-	  inputname = newvstralloc(inputname, basename, "_", number, NULL);
-	  in = fopen(inputname, "r");
-	}
-
-	if (in == NULL) {
-	  amfree(inputname);
-	  goto notincremental;
-	}
-
-	out = fopen(incrname, "w");
-	if (out == NULL) {
-	  dbprintf(("error [opening %s: %s]\n", incrname, strerror(errno)));
-	  amfree(incrname);
-	  amfree(basename);
-	  amfree(inputname);
-	  amfree(dirname);
-	  return -1;
+		dbprintf(("%s-gnutar: error opening %s: %s\n",
+			  get_pname(),
+			  inputname,
+			  strerror(save_errno)));
+		if (baselevel < 0) {
+		    goto common_exit;
+		}
+	    }
 	}
 
-	while(fgets(buf, sizeof(buf), in) != NULL)
-	  if (fputs(buf, out) == EOF) {
-	    dbprintf(("error [writing to %s: %s]\n", incrname,
-		      strerror(errno)));
-	    amfree(incrname);
-	    amfree(basename);
-	    amfree(inputname);
-	    amfree(dirname);
-	    return -1;
-	  }
+	/*
+	 * Copy the previous listed incremental file to the new one.
+	 */
+	if ((out = fopen(incrname, "w")) == NULL) {
+	    dbprintf(("opening %s: %s\n", incrname, strerror(errno)));
+	    goto common_exit;
+	}
+
+	for (; (line = agets(in)) != NULL; free(line)) {
+	    if (fputs(line, out) == EOF || putc('\n', out) == EOF) {
+		dbprintf(("writing to %s: %s\n", incrname, strerror(errno)));
+		goto common_exit;
+	    }
+	}
+	amfree(line);
 
 	if (ferror(in)) {
-	  dbprintf(("error [reading from %s: %s]\n", inputname,
-		    strerror(errno)));
-	  amfree(incrname);
-	  amfree(basename);
-	  amfree(inputname);
-	  amfree(dirname);
-	  return -1;
+	    dbprintf(("reading from %s: %s\n", inputname, strerror(errno)));
+	    goto common_exit;
 	}
-
 	if (fclose(in) == EOF) {
-	  dbprintf(("error [closing %s: %s]\n", inputname, strerror(errno)));
-	  in = NULL;
-	  amfree(incrname);
-	  amfree(basename);
-	  amfree(inputname);
-	  amfree(dirname);
-	  return -1;
+	    dbprintf(("closing %s: %s\n", inputname, strerror(errno)));
+	    in = NULL;
+	    goto common_exit;
 	}
 	in = NULL;
-
 	if (fclose(out) == EOF) {
-	  dbprintf(("error [closing %s: %s]\n", incrname, strerror(errno)));
-	  out = NULL;
-	  amfree(incrname);
-	  amfree(basename);
-	  amfree(inputname);
-	  amfree(dirname);
-	  return -1;
+	    dbprintf(("closing %s: %s\n", incrname, strerror(errno)));
+	    out = NULL;
+	    goto common_exit;
 	}
 	out = NULL;
 
 	amfree(inputname);
-      }
-      amfree(basename);
+	amfree(basename);
     }
 #endif
 
@@ -1186,39 +1178,13 @@ notincremental:
 #undef sc
 #define sc "--exclude-file="
     } else if (strncmp(exclude_spec, sc, sizeof(sc)-1)==0) {
-      efile = NULL;
-      estr = newstralloc(estr, exclude_spec+sizeof(sc)-1);
+	efile = NULL;
+	estr = newstralloc(estr, exclude_spec+sizeof(sc)-1);
 #undef sc
     } else {
-	dbprintf(("error [exclude_spec is neither --exclude-list nor --exclude-file: %s]\n", exclude_spec));
-	error("error: exclude_spec is neither --exclude-list nor --exclude-file: %s]", exclude_spec);
+	error("exclude_spec is neither --exclude-list nor --exclude-file: %s",
+	      exclude_spec);
     }
-
-    cmd_line = vstralloc(cmd,
-			 " --create",
-			 " --directory ", dirname,
-#ifdef GNUTAR_LISTED_INCREMENTAL_DIR
-			 " --listed-incremental ", incrname,
-#else
-			 " --incremental",
-			 " --newer ", dumptimestr,
-#endif
-			 " --sparse",
-			 " --one-file-system",
-#ifdef ENABLE_GNUTAR_ATIME_PRESERVE
-			 " --atime-preserve",
-#endif
-			 " --ignore-failed-read",
-			 " --totals",
-			 " --file", " /dev/null",
-			 " ", 
-			 efile ? "--exclude-from" : "",
-			 efile ? efile : "",
-			 estr ? "--exclude" : "",
-			 estr ? estr : "",
-			 NULL);
-
-    dbprintf(("%s: running \"%s\"\n", get_pname(), cmd_line));
 
     nullfd = open("/dev/null", O_RDWR);
     dumppid = pipespawn(cmd, STDERR_PIPE, &nullfd, &nullfd, &pipefd,
@@ -1228,30 +1194,34 @@ notincremental:
 	      "tar",
 #endif
 	      "--create",
+	      "--file", "/dev/null",
 	      "--directory", dirname,
+	      "--one-file-system",
 #ifdef GNUTAR_LISTED_INCREMENTAL_DIR
 	      "--listed-incremental", incrname,
 #else
-	      "--incremental", "--newer", dumptimestr,
+	      "--incremental",
+	      "--newer", dumptimestr,
 #endif
-	      "--sparse",
-	      "--one-file-system",
 #ifdef ENABLE_GNUTAR_ATIME_PRESERVE
+	      /* --atime-preserve causes gnutar to call
+	       * utime() after reading files in order to
+	       * adjust their atime.  However, utime()
+	       * updates the file's ctime, so incremental
+	       * dumps will think the file has changed. */
 	      "--atime-preserve",
 #endif
+	      "--sparse",
 	      "--ignore-failed-read",
 	      "--totals",
-	      "--file", "/dev/null",
 	      efile ? "--exclude-from" : skip_argument,
 	      efile ? efile : skip_argument,
 	      estr ? "--exclude" : skip_argument,
 	      estr ? estr : skip_argument,
 	      ".",
-	      (char *)0);
-    if (efile != NULL)
-      amfree(efile);
-    if (estr != NULL)
-      amfree(estr);
+	      NULL);
+    amfree(efile);
+    amfree(estr);
     amfree(cmd);
 
     dumpout = fdopen(pipefd,"r");
@@ -1277,14 +1247,22 @@ notincremental:
 
     kill(-dumppid, SIGTERM);
 
-    unlink(incrname);
-    amfree(incrname);
-    amfree(dirname);
-
     wait(NULL);
+
+common_exit:
+
+    if (incrname) {
+	unlink(incrname);
+    }
+    amfree(incrname);
+    amfree(basename);
+    amfree(dirname);
+    amfree(inputname);
 
     aclose(nullfd);
     afclose(dumpout);
+    afclose(in);
+    afclose(out);
 
     return size;
 }
